Improving package structure and adding back the userSetup.py version for choice in usage
